requirements.txt
aniso8601==8.0.0
Click==7.0
Flask==1.1.1
Flask-RESTful==0.3.7
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
psycopg2-binary==2.8.4
pytz==2019.3
six==1.12.0
Werkzeug==0.16.0
Flask-SQLAlchemy==2.4.1
